@@ -1,7 +1,6 @@
 'use server';
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { SecretManagerServiceClient } = require('@google-cloud/secret-manager');
+import { SecretManagerServiceClient } from '@google-cloud/secret-manager';
 const client = new SecretManagerServiceClient();
 
 export async function fetchMapTilerSecret() {
@@ -9,5 +8,9 @@ export async function fetchMapTilerSecret() {
     name: 'projects/766398966649/secrets/maptiler-key/versions/1',
   });
 
-  return secret.payload.data.toString();
+  if (secret.payload && secret.payload.data) {
+    return secret.payload.data.toString();
+  } else {
+    throw new Error('Invalid or no api key retreived');
+  }
 }
