@@ -1,5 +1,6 @@
 import React from 'react';
 
+import { SpeciesList } from './SpeciesList';
 import {
   BuoyLocationsMap,
   Card,
@@ -27,9 +28,9 @@ import {
 } from '@/utils/fns';
 import { PageProps } from '@/types';
 import { FishVariablesCard } from '@/components/visualizations/FishVariablesCard/FishVariablesCard';
-import { SpeciesList } from './SpeciesList';
 
-export default async function FishTrawl({ searchParams }: PageProps) {
+export default async function FishTrawl(props: PageProps) {
+  const searchParams = await props.searchParams;
   return (
     <React.Suspense fallback={<DefaultBuoyPage description={DESCRIPTION} />}>
       <PageWrapper params={searchParams} errorLinks={ERROR_LINKS} />
