@@ -1,7 +1,7 @@
-import { ExternalLink, Link } from '@/components';
-import { ERDDAP_URL } from '@/static/urls';
 import { HomeLink } from './HomeLink';
 import { DatasetShortcut } from './DatasetShortcut';
+import { ExternalLink, Link } from '@/components';
+import { ERDDAP_URL } from '@/static/urls';
 
 const LINK_COLORS =
   'text-white dark:text-white hover:text-teal-400 dark:hover:text-teal-300 no-underline transition-colors duraiton-500';
