@@ -1,7 +1,8 @@
 'use client';
 
 import React from 'react';
-import { Vega, VisualizationSpec } from 'react-vega';
+import { VegaEmbed } from 'react-vega';
+import type { VisualizationSpec } from 'vega-embed';
 
 import { Loading, Select } from '@/components';
 import { Size, useScreenSize } from '@/hooks/useScreenSize';
@@ -192,7 +193,7 @@ export function FishTrawlSummary({ data, options }: FishTrawlSummaryProps) {
           <Loading />
         </div>
       ) : (
-        <Vega spec={fishTrawlSpec} actions={false} />
+        <VegaEmbed spec={fishTrawlSpec} options={{ actions: false }} />
       )}
     </>
   );
