@@ -1,6 +1,7 @@
 'use client';
 import React from 'react';
-import { Vega, VisualizationSpec } from 'react-vega';
+import { VegaEmbed } from 'react-vega';
+import type { VisualizationSpec } from 'vega-embed';
 
 import type { OsomBuoyVariable, OsomSummary } from '@/utils/data/api/buoy';
 import { OSOM_VARIABLES } from '@/utils/data/api/buoy';
@@ -152,9 +153,9 @@ export function OsomBuoySummary({ data }: OsomBuoySummaryProps) {
           <Loading />
         </div>
       ) : (
-        <Vega
+        <VegaEmbed
           className="flex flex-col items-center justify-center"
-          actions={false}
+          options={{ actions: false }}
           spec={buoySummarySpec}
         />
       )}
