@@ -1,6 +1,7 @@
 'use client';
 import React from 'react';
-import { Vega, VisualizationSpec } from 'react-vega';
+import { VegaEmbed } from 'react-vega';
+import type { VisualizationSpec } from 'vega-embed';
 
 import type { RealTimeBuoyVariable, RealTimeSummaryData } from '@/utils/data/api/buoy';
 import { REAL_TIME_BUOY_VARIABLES } from '@/utils/data/api/buoy';
@@ -148,9 +149,9 @@ export function RealTimeBuoySummary({ data }: RealTimeBuoySummaryProps) {
           <Loading />
         </div>
       ) : (
-        <Vega
+        <VegaEmbed
           className="flex flex-col items-center justify-center"
-          actions={false}
+          options={{ actions: false }}
           spec={buoySummarySpec}
         />
       )}
