@@ -292,7 +292,7 @@ const ZodFetchedRealTimeBuoySummary = z.object({
   SolarIrradianceInstrumentRangeQC: z.number(),
   SolarIrradianceStuckValueQC: z.number(),
   station_name: z.string(),
-  time: z.string().datetime(),
+  time: z.iso.datetime(),
   buoyId: z.string(),
 });
 
@@ -314,6 +314,7 @@ function validateFetchedRealTimeBuoySummary(
     z.array(ZodFetchedRealTimeBuoySummary).parse(summary);
     return true;
   } catch (ex) {
+    console.error(ex);
     return false;
   }
 }
@@ -321,7 +322,7 @@ function validateFetchedRealTimeBuoySummary(
 function formatRealTimeSummaryData(fetchedData: FetchedRealTimeBuoySummary) {
   return {
     stationName: fetchedData.station_name,
-    time: new Date(fetchedData.station_name),
+    time: new Date(fetchedData.time),
     buoyId: fetchedData.buoyId,
     fluorescenceCounts: fetchedData.FluorescenceCounts,
     fdom: fetchedData.FDOM,
@@ -442,6 +443,7 @@ function validateFetchedRealTimeBuoyCoordinate(
     z.array(ZodFetchedRealTimeBuoyCoordinate).parse(coordinates);
     return true;
   } catch (ex) {
+    console.error(ex);
     return false;
   }
 }
@@ -502,6 +504,7 @@ function validateFetchedRealTimeBuoyData(
     ZodFetchedRealTimeBuoyData.parse(buoyData);
     return true;
   } catch (ex) {
+    console.error(ex);
     return false;
   }
 }
