@@ -1,8 +1,8 @@
 import { z } from 'zod';
 
-import type { FetchedFishCoordinate, FetchedTemperature, Info, SampleBase } from '@/types';
-import { getAnimalFromSpecies, getTitleFromSpecies } from '../../shared';
+import { getAnimalFromSpecies } from '../../shared';
 import { erddapAPIGet } from '../erddap';
+import type { FetchedFishCoordinate, FetchedTemperature, Info, SampleBase } from '@/types';
 
 /**
  * Fetches Coordinate information from ERDDAP.
@@ -21,6 +21,7 @@ function validateFetchedFishCoordinate(
     z.array(ZodFetchedFishCoordinate).parse(coordinates);
     return true;
   } catch (ex) {
+    console.error(ex);
     return false;
   }
 }
@@ -63,6 +64,7 @@ function validateFetchedSamples(sampleData: unknown): sampleData is { data: Samp
     ZodFetchedSamples.parse(sampleData);
     return true;
   } catch (ex) {
+    console.error(ex);
     return false;
   }
 }
@@ -106,6 +108,7 @@ function validateFetchedTemperature(temperatures: unknown[]): temperatures is Fe
     z.array(ZodFetchedTemperature).parse(temperatures);
     return true;
   } catch (ex) {
+    console.error(ex);
     return false;
   }
 }
