@@ -1,4 +1,3 @@
-import { PageProps } from '@/types';
 import {
   MA_BUOY_VARIABLES,
   MaBuoyVariable,
@@ -11,6 +10,7 @@ import {
 } from '../data/api/buoy';
 import { PLANKTON_VARIABLES, PlanktonVariable } from '../data/api/buoy/plankton';
 import { FISH_SPECIES, type FishVariable } from '../data/api/fish';
+import { PageProps } from '@/types';
 
 export type ParsedParam<T> = { error: string; value: undefined } | { error: undefined; value: T };
 type Param = Exclude<PageProps['searchParams'], undefined>[string];
